refactor: data moved to an upper level (App component)
