+++ src/App.test.tsx
@@ -1,8 +1,10 @@
-import { render, screen } from "@testing-library/react";
+import { prettyDOM, render, screen } from "@testing-library/react";
 import App from "./App";
 
 test("renders Doppler Menu Micro-Frontend", () => {
-  render(<App />);
+  const { container } = render(<App />);
+  console.log(prettyDOM(container));
+
   const el = screen.getByText("Home");
   expect(el).toBeInTheDocument();
   expect(el.classList.toString()).toBe("active");
